Working on route for adding newUserMed, still erroring

+++ src/main/java/medtrackercapstone/medtracker/database/dao/MedicationDAO.java
@@ -6,11 +6,13 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface MedicationDAO extends JpaRepository<Medication, Long> {
+public interface MedicationDAO extends JpaRepository<Medication, Integer> {
 
 //    List<Medication> findByName(@Param("name") String name);
 
     List<Medication> findAll();
+
+//    Medication findById(Integer id);
 
 
 }
